LangModelAssembly // Tests: Deprecating XCTest.
